move clean func to lib crate
