+++ lib/src/lib.rs
@@ -2,3 +2,4 @@ pub mod binary;
 pub mod env;
 pub mod image;
 pub mod io;
+pub mod path;
